pyproject.toml
[project]
name = "learnlangchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=1.4.1",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "langgraph-cli>=0.4.12",
    "onnxruntime>=1.23.2",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
